requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==4.4.2
discord.py==2.4.0
frozenlist==1.4.1
h11==0.14.0
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
multidict==6.0.5
mutagen==1.47.0
nest-asyncio==1.6.0
numpy==2.1.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
pillow==10.4.0
proglog==0.1.10
pycryptodomex==3.20.0
PySocks==1.7.1
requests==2.32.3
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
yarl==1.9.7
yt-dlp==2024.8.6
